requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
jinja2==3.1.4

# Azure SDKs (we'll use them in next step)
azure-data-tables==12.5.0
azure-servicebus
agent-framework==1.0.0b251108

# For loading env vars from .env (nice for local dev)
python-dotenv==1.0.1
python-multipart

# Teams + HTTP + Auth helpers
requests>=2.31.0
msal==1.23.0